Refactor payment processing to use double for amounts; add exampleTwo method in DesignPatternsJava to demonstrate ShoppingCart functionality with products and payment strategies

+++ DesignPatternsJava/src/strategy_pattern_package/controller/CreditCardAlgorithm.java
@@ -18,7 +18,7 @@ public class CreditCardAlgorithm implements Payment {
     }
 
     @Override
-    public void pay(int amount) {
+    public void pay(double amount) {
         System.out.println(amount + " paid with Credit Card");
     }
 }
