+++ DesignPatternsJava/src/strategy_pattern_package/controller/Payment.java
@@ -1,5 +1,5 @@
 package strategy_pattern_package.controller;
 
 public interface Payment {
-    public void pay(int amount);
+    public void pay(double amount);
 }
